bump to v27.0.0-rc.7

+++ autd3capi/src/datagram/stm/mod.rs
@@ -47,12 +47,12 @@ pub unsafe extern "C" fn AUTDSTMConfigFromPeriodNearest(p: u64, n: u16) -> Resul
 
 #[no_mangle]
 #[must_use]
-pub unsafe extern "C" fn AUTDSTMFreq(c: SamplingConfig, n: u32) -> f32 {
+pub unsafe extern "C" fn AUTDSTMFreq(c: SamplingConfig, n: u16) -> f32 {
     c.freq().hz() / n as f32
 }
 
 #[no_mangle]
 #[must_use]
-pub unsafe extern "C" fn AUTDSTMPeriod(c: SamplingConfig, n: u32) -> u64 {
-    (c.period() * n).as_nanos() as u64
+pub unsafe extern "C" fn AUTDSTMPeriod(c: SamplingConfig, n: u16) -> u64 {
+    (c.period() * n as u32).as_nanos() as u64
 }
